backend/requirements.txt
flask
flask-smorest
flask-cors
marshmallow
python-dotenv
flask-sqlalchemy
flask-migrate
psycopg2-binary
